+++ CHANGES.rst
@@ -8,6 +8,9 @@ New Features
 
 - Spinner in plot options while processing changes to contour settings. [#1794]
 
+- Model fitting plugin can optionally expose the residuals as an additional data collection entry.
+  [#1864]
+
 Cubeviz
 ^^^^^^^